LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/at-property-expected.txt


FAIL Attribute 'syntax' returns expected value for ["<color>"] undefined is not an object (evaluating 'rule.syntax')
FAIL Attribute 'syntax' returns expected value for ["<color> | none"] undefined is not an object (evaluating 'rule.syntax')
FAIL Attribute 'syntax' returns expected value for ["<color># | <image> | none"] undefined is not an object (evaluating 'rule.syntax')
FAIL Attribute 'syntax' returns expected value for ["foo | bar | baz"] undefined is not an object (evaluating 'rule.syntax')
FAIL Attribute 'syntax' returns expected value for ["*"] undefined is not an object (evaluating 'rule.syntax')
FAIL Attribute 'syntax' returns expected value for ["notasyntax"] undefined is not an object (evaluating 'rule.syntax')
FAIL Attribute 'syntax' returns expected value for [red] undefined is not an object (evaluating 'rule.syntax')
FAIL Attribute 'syntax' returns expected value for [rgb(255, 0, 0)] undefined is not an object (evaluating 'rule.syntax')
FAIL Attribute 'syntax' returns expected value for [<color>] undefined is not an object (evaluating 'rule.syntax')
FAIL Attribute 'syntax' returns expected value for [foo | bar] undefined is not an object (evaluating 'rule.syntax')
FAIL Attribute 'initial-value' returns expected value for [10px] undefined is not an object (evaluating 'rule.initialValue')
FAIL Attribute 'initial-value' returns expected value for [rgb(1, 2, 3)] undefined is not an object (evaluating 'rule.initialValue')
FAIL Attribute 'initial-value' returns expected value for [red] undefined is not an object (evaluating 'rule.initialValue')
FAIL Attribute 'initial-value' returns expected value for [foo] undefined is not an object (evaluating 'rule.initialValue')
FAIL Attribute 'initial-value' returns expected value for [if(){}] undefined is not an object (evaluating 'rule.initialValue')
FAIL Attribute 'initial-value' returns expected value for [var(--x)] undefined is not an object (evaluating 'rule.initialValue')
FAIL Attribute 'inherits' returns expected value for [true] undefined is not an object (evaluating 'rule.inherits')
FAIL Attribute 'inherits' returns expected value for [false] undefined is not an object (evaluating 'rule.inherits')
FAIL Attribute 'inherits' returns expected value for [none] undefined is not an object (evaluating 'rule.inherits')
FAIL Attribute 'inherits' returns expected value for [0] undefined is not an object (evaluating 'rule.inherits')
FAIL Attribute 'inherits' returns expected value for [1] undefined is not an object (evaluating 'rule.inherits')
FAIL Attribute 'inherits' returns expected value for ["true"] undefined is not an object (evaluating 'rule.inherits')
FAIL Attribute 'inherits' returns expected value for ["false"] undefined is not an object (evaluating 'rule.inherits')
FAIL Attribute 'inherits' returns expected value for [calc(0)] undefined is not an object (evaluating 'rule.inherits')
PASS Invalid property name does not parse [foo] 
PASS Invalid property name does not parse [-foo] 
FAIL Rule applied [*, if(){}, false] assert_equals: expected "if(){}" but got ""
FAIL Rule applied [<angle>, 42deg, false] assert_equals: expected "42deg" but got ""
FAIL Rule applied [<angle>, 1turn, false] assert_equals: expected "360deg" but got ""
FAIL Rule applied [<color>, green, false] assert_equals: expected "rgb(0, 128, 0)" but got ""
FAIL Rule applied [<color>, rgb(1, 2, 3), false] assert_equals: expected "rgb(1, 2, 3)" but got ""
FAIL Rule applied [<image>, url("http://a/"), false] assert_equals: expected "url(\"http://a/\")" but got ""
FAIL Rule applied [<integer>, 5, false] assert_equals: expected "5" but got ""
FAIL Rule applied [<length-percentage>, 10px, false] assert_equals: expected "10px" but got ""
FAIL Rule applied [<length-percentage>, 10%, false] assert_equals: expected "10%" but got ""
FAIL Rule applied [<length-percentage>, calc(10% + 10px), false] assert_equals: expected "calc(10% + 10px)" but got ""
FAIL Rule applied [<length>, 10px, false] assert_equals: expected "10px" but got ""
FAIL Rule applied [<number>, 2.5, false] assert_equals: expected "2.5" but got ""
FAIL Rule applied [<percentage>, 10%, false] assert_equals: expected "10%" but got ""
FAIL Rule applied [<resolution>, 50dppx, false] assert_equals: expected "50dppx" but got ""
FAIL Rule applied [<resolution>, 96dpi, false] assert_equals: expected "1dppx" but got ""
FAIL Rule applied [<time>, 10s, false] assert_equals: expected "10s" but got ""
FAIL Rule applied [<time>, 1000ms, false] assert_equals: expected "1s" but got ""
FAIL Rule applied [<transform-function>, rotateX(0deg), false] assert_equals: expected "rotateX(0deg)" but got ""
FAIL Rule applied [<transform-list>, rotateX(0deg), false] assert_equals: expected "rotateX(0deg)" but got ""
FAIL Rule applied [<transform-list>, rotateX(0deg) translateX(10px), false] assert_equals: expected "rotateX(0deg) translateX(10px)" but got ""
FAIL Rule applied [<url>, url("http://a/"), false] assert_equals: expected "url(\"http://a/\")" but got ""
FAIL Rule applied [<color>, tomato, false] assert_equals: expected "rgb(255, 99, 71)" but got ""
FAIL Rule applied [<color>, tomato, true] assert_equals: expected "rgb(255, 99, 71)" but got ""
PASS Rule applied for "*", even with no initial value 
PASS Rule not applied [undefined, green, false] 
PASS Rule not applied [<color>, undefined, false] 
PASS Rule not applied [<color>, green, undefined] 
PASS Rule not applied [<gandalf>, grey, false] 
PASS Rule not applied [gandalf, grey, false] 
PASS Rule not applied [<color>, notacolor, false] 
PASS Rule not applied [<length>, 10em, false] 
FAIL Non-inherited properties do not inherit assert_equals: expected "40px" but got " 40px"
FAIL Inherited properties inherit assert_equals: expected "40px" but got " 40px"
FAIL Initial values substituted as computed value assert_equals: expected "rgb(0, 128, 0)" but got ""
PASS Non-universal registration are invalid without an initial value 
FAIL Initial value may be omitted for universal registration assert_equals: expected "" but got " calc(1px + 1px)"
